UPBGE: Implement custom filter off screen.
Previously all the filter were rendering to a game engine off screen used as
input for the next filter.
But in some cases the user will need to temporary store the result of a filter
to reuse it in multiple filters for multiple reasons:
- Need to use the same data for two filters.
- Store computation result to optimize filter rendering.

To do so we have to allow the user to render to a custom off screen which will
store the temporary data. This off screen is unique per filter.

In the python API side, the user create the off screen calling:
KX_2DFilter.addOffScreen(slots, depth, width, height, hdr, mipmap)

All the variable are explicit excepted slots, width and height. slot is the
number of color textures attached to the off screen. width and height have the
particularity to let the off screen be resized to the viewport size if there's
value is to -1.

The created off screen is an instance of KX_2DFilterOffScreen and accessible
from KX_2DFilter.offScreen.

After the off screen creation the user could specify where he writes in the
filter shader using gl_FragData. Once the off screen was filled its attached
textures could be reused in the next filter shader by accessing the color
texture bind code:
KX_2DFilterOffScreen.colorBindCodes[x].

This bind code is set using KX_2DFilter.setTexture(…).

If the user want to disable the render to this custom off screen, he could
remove it calling KX_2DFilter.removeOffScreen().

In the internal source side the thinks are pretty similar, the filter off screen
comes from RAS_2DFilterOffScreen overridden into KX_2DFilterOffScreen for the
python API.

This class mainly manages the creation of the frame buffer and of the attached
textures. These textures are created at the filter off screen creation if the
width and height are fixed, else the textures can be created later when the
filter do its first render or when the viewport dimensions changed. This dynamic
update is proceed in RAS_2DFilterOffScreen::Update(canvas).

As the filter were able to render to a custom off screen the targeted off screen
of a filter is not predictable from the RAS_2DFilter::Start call. For this
reason this same function return the off screen to use as the input off screen
of the ext filter. This returned off screen can be the targeted off screen sent
as parameter to Start if there's no custom off screen or the input off screen
sent to Start if the there's a custom off screen.

Inside RAS_2DFilter::Start, we check if there's a custom off screen and if it
is valid, if so, the off screen is bound and the viewport is changed in case of
fixed off screen size. After this the filter render is proceeded and finally
the custom off screen is unbound (if it exists) and the mipmap could be
generated.

Following this behaviour we could not ensure that the returned off screen of
Start is currently bound as before. Indeed, in case of custom off screen
render only the custom off screen is yet bound. For this reason the last filter
must not use a custom off screen.