UPBGE: Fix for filterManager.getFilter warning message
Previously, a message was displayed when getFilter method returned None.
This behavior is not correct because we have to keep the possibility to
write: if scene.filterManager.getFilter(0) is not None: do something
without having a message splatted in the console.

So the warning is removed from the getFilter method and added in the
KX_2DFilterManager API documentation