UPBGE: Refactor scene converter.
Previously the converter was using to store the data of the scenes and the links
to them during a scene conversion. These two goals could be split into two classes
to avoid update the links and clear the code by avoiding access data that are dangling.

Also in case of two lib loading in asynchronous the links are written and read by
two conversion task at the same time, which involved unstable behavior.

To avoid the issues described before the scene converter is distinguished of the
scene data store. The converter keep is name as KX_BlenderSceneConverter but its goal
is restricted.

The scene converter is temporary for the scene conversion, it is passed as reference
to ensure to not store a pointer to it. It contains the map between the blender data
and the game engine data like: objects, meshes, material, actuators and controllers.
To access and register these data the functions follow the patterns:
RegisterGEDataType(BlenderType, GEDataType)
FindGEDataType(BlenderType)

The scene converter also store a list of the game engine data types registered during
the scene conversion. These are poly material and mesh objects.

Once the conversion is over, the scene converter is deleted but the list of materials
and meshes in it are copied into the KX_BlenderConverter.

This new class has the goal of launching the scene conversion and storing data slots
per scenes. These slots contains approximately the same data than the list stored in
the scene converter excepted for the action interpolators which are converted dynamically
after the scene conversion.

These scene slots own the materials and meshes in they. A scene slot can be created
using a scene converter or merged with an other scene slot in case of scene lib loading
conversion.

Other than that the new KX_BlenderConverter is keeping the functions in the previous
KX_BlenderSceneConverter without the one used for data linking.

Note: MSVC doesn't support using a type without destructor in a std::map even if the
destructor are all defined in .cpp, to solve this issue we need to add extra includes.