UPBGE: Remove auto resize in image_new_exec for environment maps.
The auto resize is remove because at this moment we don't know if the
user is wanting a cube map or a planar map and the second doesn't
need a proper size.

The auto resize is removed but a invalid message/draw in UI will
warn the user that the size are incorrect in the case of cube maps.