UPBGE: Fix replication of armature constraint objects.
The objects used for the armature constraints are blender objects, but previously
when an armature was replicated the blender objects weren't, so the solving of
the constraints was wrong.

To solve this issue we need to use a new blender object. Instead of replicating
the original blender object targeted, we simply use a dummy object which is
an enpty blender object created by each BL_ArmatureConstraint, this object is
set to the constraints calling flush_constraint_targets with no_copy to 0.

For constraint targeting an other armature bone, the pose is set into the blender
object in BL_ArmatureConstraint::UpdateTarget.

These object are then freed at the BL_ArmatureContraint desctruction with a pose
set to nullptr to avoid free a not owned armature pose.

Fix issue #393.