UPBGE: Begin small text refactor
We need to disable lighting to draw 2D texts. We don't need to process
lights neither for 2D, neither for 3D texts
(lighting and shadowing are not supported during runtime neither on BF
blender neither on upbge master). Lights and shadows are only supported in
the viewport.

We have to keep padded text for profile because of profile visibility but
we don't need the enum.

About transparency, only 3D texts needs it so I moved call to
SetAlphaBlend(GPU_BLEND_ALPHA) in RenderText3D.