UPBGE: Fix replication of bounding box.
Previously when a RAS_BoundingBow was duplicated from a deformer,
it wasn't added back into the bounding box manager, so this bounding
box was able to be active but not owned by the manager.
This behaviour without libload caused only memory leaks, but in the
case of a libload the bounding box manager are merged and each bounding
box switch of manager pointer, but as the duplicated bounding box were
not know by the manager merged, they then didn't change of manager pointer.
At the release of all the users of a bounding box, the bounding box ask
its manager to remove it from the active bounding box list. Here the
crash happen on a dangling manager pointer.

To solve this bug we MUST add the bounding box into the manager when it
is duplicated. In the same time the RAS_BoundingBoxManager is a friend
class of RAS_BoundingBox so it can add itself into the manager without
calling a special function and restricting this operation only to this
class.

Fix a part of #382.