UPBGE: Temporarly remove parallax uv shift option
This option was already disabled in
https://github.com/UPBGE/blender/commit/686112da3bdb527447933fbf1fa0f7684ae433ca
This leads to confusion for the user:
https://github.com/UPBGE/blender/issues/354

There was a real problem with naming in the code:
- float scale in the shader (not explicit)
- heightscale in the UI (not explicit)
- parallax uv or parallax uv shift (last name is the most explicit name
  for me)

I think if we want such an option we have to allow shifting on U and V
axis.

Waiting we find a new solution, I propose to remove this from the UI and
the code.