UPBGE: Fix wireframe color.
Previously the wireframe render was interfered with the color used previously.
This color was set from the render of the debug lines and not unset before
render as wireframe.

In place of override the color, a shader is used to make sure that the color
will never depends of a correct or wrong state.
This shader is named the black shader and accesible at the enumerations:
GPU_SHADER_BLACK
GPU_SHADER_BLACK_INSTANCING

They are bind at the place of the old "basic" shader into RAS_BucketManager.

Bug reported by TwisterGE.