UPBGE: Fix buffer allocation in IMB_allocFromBuffer.
In IMB_allocFromBuffer a copy of the image data is proceeded with a call to
MEM_dupallocN. But all the MEM_ prefix functions work with memory allocated
with the same function.
This was not the case in Image::init, so the copy failed and the all the
operations later used to mipmap were invalid and raise segmentation faults.

To fix this issue we use MEM_mallocN in Image::init.

Fix issue #411.