UPBGE: Fix loading of font for object in linked libraries.
Previously the loading of the font was using the library path (exactly the
main path) of the first library converted, but in case of lib loading or
linked library the path was invalid. To fix this issue we use the id.lib->filepath
from the vfont passed to the function GetFont.

Fix issue #377.