UPBGE: Fix string to float conversions execptions.
With the switch to std::string the previous conversion made with STD_String::ToFloat
were proceeded with std::atof now. But this function raise an exception when the
conversion failed.
The alternative is to use std::stringstream which do a safe conversion and have
a stet to check the succes of the conversion. But unfortunatly its use request a
boring syntax. To avoid this a template function is written in CM_Format.h.
This function is named CM_StringTo and two argument: the string and a reference to
the type to convert to.

This new function is now used in the property sensor for string to float converions.