UPBGE: Avoid merge object of a batch group into an other batch group.
Previously this behaviour was totally allowed but it is not aprecated.
Firstly it can hide a recursion merge which could do a high usage of
memory.
Secondly merging a object already used in a batch group to an other
batch group is like merging the both group, but if the first batch group
split or merge one object, the compound ot the batch group will be not
update and very hard to update.

Fix issue #416.