UPBGE: Fix OpenGL state before drawing texts.
Previously a function named DisableForText was used befoe rendering a
text in 3D or 2D or a 2D box. But in the refactor of the rasterizer
and the usage of the pimpl idiom, the function was split in two
function in the pimpl class (RAS_OpenGLRasterizer) and the rasterizer
(RAS_IRasterizer).

Unfortunatly only the pimpl class function was called into the pimpl
without calling the rasterizer function. This caused to not set a proper
OpenGL state before rendering the texts.

To fix this issue the rasterizer function is properly called from the
pimpl class.

Also the DisableForText was calling ProcessLighting to disable the lights
but using a call to DisableLights is better because it doesn't request
the last client object pointer and don't do useless check of last light
layer.

Fix issue #396.