UPBGE: Remove /bigobj option that caused issue on MSVC 2013
When we compiled upbge with MSVC 2015 we needed /bigobj option to build in
debug mode else we had compilation errors.

But this has a bad effect on MSCV 2013:
http://pasteall.org/pic/show.php?id=113567 Debug information format was
empty. When this option is empty, we can't use breakpoint to debug in MSVC
2013.

For now I manually set it wit /Zi (the correct option). But as we build
upbge with MSVC 2013 and that this commit that added /bigobj flag had this
indesirable effect, I remove this commit for now because we don't need it
on MSVC 2013.

If we want to do the things correctly, we have to restore /bigobj only
for MSVC 2015 (if (MSVC14) in cmake) and find a way to fill debug
informations format field with /Zi.