UPBGE: Refactor debug draw options.
Previously some of the debug options were global as the bounding box or physics
and also per objects like the armatures and cameras. But in the two situations
the user is bored: if the user want a per object setting instead of a global
option and if the user want to enable or disable the debugging for all the objects
instead of disabling it with a per object option.

To make the debug options easier three mode are used:
- Force: force to draw the debug draw for every objects.
- Disable: don't do any debug draw.
- Allow: let per object settings choose to draw the debug.

These three options are implemented for the debug option for bounding box, armatures
and cameras, this last option was implemented and is added in this commit.

In the RNA sources, these properties are defined with a enum, 0 for disable, 1 for
enable and 2 for allow.
In the python UI, the option are set in a special panel dedicated only for debug
options and a sub menu for the game menu is made for the debug option. This menu
show the enumeration as a list of check options as the other enumerations in the
UI menus.
As the debug option are removed from display panel, only animation fps and frame
color were left, the animation fps is pretty unrelated to the display and is moved
into its own panel with the restrictive animation update option.

In the game engine sources, the command line option "show_camera_frustum" is added.
Also the option are stored by the C++ enum class named KX_DebugOption.