UPBGE: Use enum class for KX_ExitRequest.
Previously a normal enumeration were used but each enumeration
values were using a prefix in theirs names. To avoid this we
use a enum class (since C++11) to set the name of the enumeration
used as the prefix of each values. An enum class also have the
advantages to use a strict type without implicit casts.