UPBGE: Refactor scene suspended time managment.
Previously when a scene was suspended, the function UpdateSuspendedScenes
was checking that the suspended scene has a valid suspended time, this
value is the clock time when the scene was suspended. If this value was
to zero it was set.
After, when the scene was resumed, the same function checked if the
suspended time was not zero, in this case the scene was just resumed, the
delta suspended time (the time passed during the scene suspend) wad set
and the suspended time was set to zero.

This mechanism is to complicated for the informations we need, indeed
only the suspended delta time is used for animations. This delta time
could be incremented each frame by the frame step.

So the suspended time of a scene is removed and UpdateSuspendedScenes get
as argument the frame step which is incremented to the scene suspended
delta time.

Tested with #D1569.