UPBGE: Don't use pointer for KX_Time[Category]Logger.
KX_TimeCategoryLogger was owned by KX_KetsjiEngine and KX_TimeLogger
was owned by KX_TimeCategoryLogger. But they were not accesable outside
their class owner.
In this case the usage of a pointer is weak and restrict a more
boring syntax and allocation management.
Using by value instead of pointer avoid this complex syntax and remove
the manually allocation.