UPBGE: Refactor non bullet culling management.
The code managing the culling without bullet (no occlusion) was cleaned and
simplified. Firstly the classes used in the culling are explicitly defined.
Three classes are added:

- SG_CullingNode : The node owned by an object which can be culled, it contains
an AABB. This class is inherited by KX_CullingNode, this final class is created
by every KX_GameObject and store a pointer to its owner object.
- SG_Frustum : The frustum of a camera, containing the camera frustum matrix and
the pre-computed planes.
- KX_CullingHandler : The handler of the culling for a scene. This class is
temporary and accept a frustum and a list reference of culling nodes, after this
every object culling node in the scene go through the handler which each select
the nodes to push in the list.

As explained upper, the culling nodes are put into a list when they are not
culled during the culling pass, this list is reused by the functions updating
the lod, or drawing the meshes. These functions benefits of the optimization to
iterate on the non-culled objects instead of iterating the scene object and
checking the culling state.

The culling test algorithms are also refactored and fixed. The old box or AABB
checking was doing false positive in case the box was to large and then
intersecting two "orthogonal" frustum plane and not inside the frustum volume.
This issue is fixed by totally changing the way we check an AABB (here box
became a real AABB). The function doing this test is
SG_Frustum::AabbInsideFrustum. This keep the AABB and transform the frustum
plane into the object space, each planes are testing the closest and further
AABB point computed with the axis of the plane normal. If the closest point is
out the AABB is out, else if the further point is out the the AABB intersect the
frustum. Else if the closest and further points are always in all the planes,
the AABB is inside the frustum.
At this point the false positives are still presents, to avoid them the AABB
should be in the frustum volume AABB, in all other cases the object AABB is
either out of plane or out of the frustum AABB.

All the other volume frustum tests are moved into SG_Frustum, excepted for
sphere test which use a method similar to the point test which allow to not
computing a frustum center and radius.

The culling time spent when all the objects are visible is mostly the same, but
when all the objects are invisible the render time pass from 0.23 ms to 0.08 ms
for 8000 objects. Indeed the function KX_GameObject::UpdateBuckets is not called
for culled objects.