UPBGE: Implement CListValue filtering by name and properties.
This commit implement a method named filter in CListValue
which take two arguments name and prop. These two arguments
are used in a regex search of the list item which have a
matching name or a matching property.

ex:

scene.object.filter("C.*", "Enemy")

Will search every object with name starting by C and with a
property named Enemy.