UPBGE: Implement font object culling bounding box.
Previously the font objects wasn't doing any culling computation they there
always visible. This behaviour have three desavantages:
- the render can't be optimised with culled font object
- the sources request to indentify font objects at rendering
  stage to don't care about the culling state
- the culling state from python is totally unusable.

To solves all the issues the best way is to implement a computation of the
bounding box for the fonts. This computation is made in the function
KX_FontObject::GetTextAabb. This function get the text AABB using BLF
function BLF_boundbox with the bound settings for dpi and size.

To make sure the AABB is computed in the wanted cases, the modification
of the text is done calling the function SetText. This function set the
text and compute the AABB.

One of the encountered issue was that the font objects could use a property
name "Text", but this property was read in UpdateBucket just before the
render and the AABB need to be computed before.
To avoid this issue the scene is iterating over all the font objects at the
logic frame end and calling the function KX_FontObject::UpdateTextFromProperty.
This function compare the current text with the text in a possible property
named "Text" and copy (and compute AABB so) if they are different. To compare
the font object have to store the full text into a full string and a array of
string per lines, this is why SetText use as argument a full string instead
of an array.

In the same time a bug with KX_FontObject.dimensions in python was noticed.
Indeed this function was based on a BLF function returing the size of a
single string, this technique totally failed for a multilines font object
even applying a line spaccing or something else.
The solution was to get the width and height of the bounding box.

Previously the culling pass was checking that the object were using a mesh
before process theirs culling, but this fails for font objects. Instead
checking the number of mesh we call the function KX_GameObject::UseCulling.