UPBGE: Fix conversion of node materials.
The node materials could use other materials in is node tree. So these
materials could be not used directly by faces of a mesh.

This last behaviour was causing issues because previously the mesh
conversion was doing the material conversion of only materials found
into faces.
So materials used by a node tree could be not converted. This caused
also that the texture rendered were not converted in the bug report

To fix this, all the materials are converted before converting the faces
and then get using RAS_MeshObject::GetMeshMaterialBlenderIndex. Else
the conversion of the uv and color layer are made by the new function
uvsRgbFromMesh, in consideration material_from_mesh is doing only
the material conversion without uv and color layers.

Fix issue #450.