UPBGE: Cleanup shadow shader and implement slope bias.
The shader code managing shadow is clenaup by separating the computation
in different functions reused by different shadow strategie. For example
the texture coordinate is computed by the function shadow_proj_coord and
reused for simple shadow, PCF and VSM. The same is made for shadow_visibility.

Each shadow technic use two function, the first is prefixed test_shadow_[technic]
and simply return a float coresponding to the shadow on the fragment,
we pass to this function the texture coordinate directly.
The second function is used during material generation and is named
shadow_[technic].

In the second time, slope bias is implemented. This technic is used to
avoid shadow acne by not using only a constant bias corresponding to depth
shift, but a shift depending on the geometry tested. The best technic is
to use as fragment position for shadow the real fragment position combined
with the geometry normal multiplied by a factor. The final fragment
position is multiplied by shadow matrix and the shadow test is proceeded
normally.