UPBGE: Cleanup option and flag management in KX_KetsjiEngine.
Too many options were used into KX_KetsjiEngine and a lot more will
be added later. To avoid mess the code with getters/setters and members
an enum is used with a common function called SetFlag. This function
receive the flag type to modify and the state to set.
Once this the flag could be read using the getter GetFlag.

In the same time the unused attributes in KX_KetsjiEngine were removed.