UPBGE: Expose time scale in UI.
The time scale is used to speed up or slow down animations and physics.
A new option is added in the Scene panel under Physics with the name
"Time Scale".

The value of the time scale is copied into the ketsji engine at its
creation in the launcher. The same is made for attributes previously
copied in KX_KetsjiEngine::StartEngine().