UPBGE: Implement CM_RefCount.
This class unified the behaviour of the classes using a reference counter.
It exposes three functions, AddRef, Release and GetRefCount.
Thanks to CRTP the functions AddRef and Release can return a pointer to
the object of the right type, but this doesn't work with an inheritance of
multiple levels. To solve this issue helper functions CM_AddRef and CM_Release
are used, they resepectivly call AddRef and Release and return the proper
type pointer using a cast of the object type passed as argument.

CM_RefCount is used for CValue, CExpression, KX_LodManager,
RAS_DisplayArrayBucket and CcdShapeConstructionInfo.