UPBGE: Refactor stereo render management.
Previously the cameras data where computed just before the real render, so we
were able to modify this computation in case of stereo as did the function
RAS_Rasterizer::SetRenderArea which depended of the current stereo eye.

But with the recent work about the culling of shadow or texture probes we need
to get the camera data before any render. The easiest way is to force the
computation of the modelview and projection matrix before the render, but this
not work for stereo because we used twice the same camera.
To solve this issue we have to duplicate the camera in two new camera in case of
stereo, and these camera have to be stored in a structure to be able to be freed
at the render end.

In the same time computing the needed camera before the render could avoid do it
again during the render, then we need to store the result in a structure
hierarchy.

This hierarchy is made of three structure owning each other in a kind of tree:
- FrameRenderData: represents the render to a category of off screen, two are
used in stereo modes using a separate off screen, else only one is used.
- SceneFrameRenderData: represents the render of a scene, a scene could not have
an instance if it doesn't have a valid camera.
- CameraFrameRenderData: represents the render of a camera, contains the camera
to render its viewport and its eye value (left by default).

This render frame data tree is generated in the function
KX_KetsjiEngine::GetFrameRenderData which is called into
KX_KetsjiEngine::Render(), once we get the render frame data, we iterate over
all frame and select/bind the appropriate off screen, then we iterate over the
scenes and update world settings, then we iterate over all camera and render a
part of the frame.

As the stereo is managed the same way that a normal render, the exceptions made
in KX_KetsjiEngine::Render fro per eyes stereo could be removed and then the
usage of the old RAS_OFFSCREEN_RENDER/FINAL is the same than
RAS_OFFSCREEN_LEFT_EYE0/1 if we take the left eye as the default off screen of a
normal render. So two off screens are removed.

Because the computation of the camera data doesn't depends of the current eye
state in the rasterizer the function RAS_Rasterizer::SetRenderArea became
GetRenderArea which returns a RAS_Rect and use as argument a eye value. The same
is made for GetFrustumMatrix and the new GetViewMatrix.

Note: the python callback KX_Scene.pre_draw_setup in case of stereo receive a
temporary camera and all the camera are passed through this callback before the
other callback KX_Scene.[pre/post]_draw.