UPBGE: Fix negative text size.
Previously when a text where using a negative scale a huge amount of memory
were used, this was caused by the fact that we managed the size as a signed
int but the BLF modules managed it as an unsigned int, so when the size was
negative BLF used a value close to 2^32 as size.

To solve this issue we use an absolute size into KX_FontObject, in the same
time the negative scale mirror render is fixed.