UPBGE: Fix bullet3d culling with inactive objects.
Calling SetVisible in a inactive object has the side effect to add its
graphic controller into the culling tree when bullet is used for the
culling. This as the effect to make a inactive object non-culled,
this is not good but the easiest way instead of avoiding adding the
controller and then add the controller at the replication etc…

Previously that wasn't important because we iterate over all the objects
in the rendering and rendered the non-culled ones, but now as we use
a list of objects we can't know if the objects are all active.

To avoid this situation the function PhysicsCullingCallback which selects
if an object is culled or not, make sure that the object can be culled.