add logic culling
The technique used is to set the object state (that correspond to
controller state) to an inexistant state with no logic bricks. Like that
the logic is suspended. We don't have to iter on all controllers to
remove them. It's easy to suspend and easy to restore and it costs no
performance.