UPBGE: Implement intersection test of frustum volumes.
The intersection of two frustum volumes is based on the Separate
Axis Theorem (SAT). This theorem proof that if two objects don't
intersect then it exists a plane there the projection of the objects
is disjoint. These planes are the one formed by the normal of each
objects and the one based on the cross product of a pair of edge
of the objects.

https://booksite.elsevier.com/9781558605930/revisionnotes/MethodOfSeperatingAxes.pdf