UPBGE: Allow last filter to render to its off screen.
Previously the last filter wasn't able to render to its custom
off screen because we rely on the filter to do the copy to our
internal target off screen. This behaviour is avoid by comparing
the last used off screen at the end of the filter pass with the
normal target off screen, if they differs a simple draw is proceed.