UPBGE: Fix add of armatures in animated object list.
Previously the armature where's added to the animated object list by
calling KX_GameObject::GetActionmanager in BL_ArmatureObject and
replication function. But in the case of an armature contained in a
dupli group from a libloaded scene, the scene of this object is not
yet scene and GetActionManager which use internally GetScene will
fall back to add the armature into the scene calling the conversion,
not the one converted.

To fix this issue GetActionManager should always use the object
scene not the global active scene, to do this KX_GameOjbect::GetScene
always return the scene graph node scene.

Secondly the armature object should be added to the animated object
list without calling GetActionManager, so AddNodeReplicaObject
is checking for armature type and adding if true. The same is proceeded
at the scene conversion.

Finally KX_Scene use a std::vector instead of CListValue because the
list isn't exposed in python and that it is better to manage double
insertion in AddAnimatedObject than with CListValue. In the same time
no reference counting is used with this list.

Fix issue #478 (first bug).