UPBGE: Fix duplication of display array bucket with shared storage info.
Previously when a display array bucket was duplicated the storage info
owned in it wasn't set to nullptr to be recreated at the next render.
This caused to disallow armature to modify the VBO and to double free
the storage info as they are supposed to be only owned by the DAB.

To fix this issue the storage info is set to nullptr in
RAS_DisplayArrayBucket::ProcessReplica to request to recreate it during
the next render.

Fix issue #466.