UPBGE: Fix stereo camera reference count.
In case of stereo the camera used to render are duplicated for
each eyes so way need an extra reference decrease after the their
creation.
With the introduction of GetCameraRenderData the call to Release
was made before returning the camera render data which freed the
camera before its use.
To fix this issue the Release is made after creating the camera
render data.