UPBGE: Fix libloading of 2d filter actuator.
Previously the 2d filter actuator was using internally a pointer
to the filter manager of the scene owning the actuator. But in case
of libloading ths pointer wasn't updated.

To fix this issue the function SCA82DFilterActuator::SetScene now
takes as second arguments the filter manager of the new scene owning
the actuator during the logic brick merge.

Fix issue #488.