UPBGE: Improve tooltip for margin in physics tab
The default margin should always be 0.04. The default Cube has a margin of
0.06. This is only a bit annoying if you set this cube to rigid body,
enable collision bounds with triangle mesh, add a plane with margin of
0.04 and set the plane to static with triangle mesh because it creates a
small jitter.

This jitter doesn't occur when plane and cube have their margin set to
0.04.

But:
1) I didn't found how to change margin for default cube (it doesn't work
in versioning legacy)
2) When you add an object, its default margin is set to 0.04
3) This seems a rare bug

So I just propose to improve tooltip of margin with the following patch.

I think this is enough to answer this report:
https://github.com/UPBGE/blender/issues/484